api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.20
mysql-connector-python>=8.1.0
pydantic>=2.10.0
python-dotenv>=1.0.0
httpx>=0.27.0
google-ads>=24.0.0
